+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.0.0-rc.7 - 2018-07-09
 ### Fixed
 - Ensure a string is returned when calling ObjectFromElementAccessor::getId
 - ObjectsField::normalizeQueryInputValue should return a record, not a query